requirements.txt
opencv-python==4.5.5.64
torch
torchvision
torchaudio
tensorflow
scikit-learn==1.5.0
matplotlib==3.7.1
seaborn==0.12.2 
numpy==1.23.5
pandas==1.5.3
tqdm==4.66.3
mtcnn==0.1.1
facenet-pytorch==2.5.3
streamlit==1.37.0
Pillow
keras
keras-facenet==0.3.2